Re-enable lint to ensure no debug code loaded into extension
