+++ lints/disallowImportingNonSharedCodeRule.js
@@ -14,10 +14,7 @@ Rule.EXTENSION_FAILURE_STRING = "Do not import extension code because the extens
 
 class NoNonSharedCode extends Lint.RuleWalker {
 	visitImportDeclaration(node) {
-		// TODO: Remove first part of this condition when DAs are not running in process.
-		// https://github.com/Dart-Code/Dart-Code/issues/1876
-		if (this.sourceFile.fileName.indexOf("debug_config_provider") === -1
-			&& node.moduleSpecifier.text.indexOf("../debug/") !== -1) {
+		if (node.moduleSpecifier.text.indexOf("../debug/") !== -1) {
 			this.addFailure(this.createFailure(node.getStart(), node.getWidth(), Rule.DEBUG_FAILURE_STRING));
 		}
